Move hardware and firmware data source classes
